rest/requirements.txt
ciso8601==2.1.3
python-dotenv==0.13.0
requests==2.23.0
